WIP potential fix using generator
